+++ frontend/src/App.jsx
@@ -1,13 +1,15 @@
 import "./components/css/App.css";
 import { Routes, Route } from "react-router-dom";
-import Home from "./components/Home";
-import Appointment from "./components/Appointment";
+import Home from "./components/pages/Home";
+import Login from "./components/pages/Login";
+import Signin from "./components/pages/Signin";
 
 function App() {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
-      <Route path="/appointment" element={<Appointment />} />
+      <Route path="/login" element={<Login />} />
+      <Route path="/signin" element={<Signin />} />
     </Routes>
   );
 }
