--- frontend/src/components/Appointment.jsx
@@ -1,7 +1,0 @@
-export default function Appointment() {
-  return (
-    <div>
-      <h1>Voici la page de réservation, félicitation !! 😁</h1>
-    </div>
-  )
-}
