+++ frontend/src/components/css/App.css
@@ -56,7 +56,7 @@ h3 {
   src: url(../../../public/assets/font-family/RobotoSlab-Regular.ttf);
 }
 
-.bg-grey {
+.bg-gray {
   background-color: #1c1e22;
 }
